extract DataPropertyInterpreter from IginxInterpreter8
